+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/trip/model/RouteProgress.kt
@@ -41,7 +41,7 @@ import com.mapbox.navigation.base.trip.model.roadobject.UpcomingRoadObject
  * this field can provide the route ID of an alternative route that user turned into causing off-route event (if there is one).
  * This field can be used to find a route with [NavigationRoute.id] that can be immediately used as the new primary route.
  * @param currentRouteGeometryIndex route-wise index representing the geometry point
- * right in front of the user (see [DirectionsRoute.geometry]), null if unavailable.
+ * right in front of the user (see [DirectionsRoute.geometry]).
  */
 class RouteProgress internal constructor(
     val navigationRoute: NavigationRoute,
@@ -59,7 +59,7 @@ class RouteProgress internal constructor(
     val upcomingRoadObjects: List<UpcomingRoadObject>,
     val stale: Boolean,
     val routeAlternativeId: String?,
-    val currentRouteGeometryIndex: Int?,
+    val currentRouteGeometryIndex: Int,
 ) {
 
     /**